# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventPhotoRequirement = typeof EventPhotoRequirement[keyof typeof EventPhotoRequirement];


export const EventPhotoRequirement = {
  none: 'none',
  optional: 'optional',
  required: 'required',
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  /** Time string (HH:MM) */
  time: string;
  /** @nullable */
  dresscode?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  registrationUrl?: string | null;
  /** @nullable */
  tickettailorEmbed?: string | null;
  /** @nullable */
  areaDescription?: string | null;
  /** @nullable */
  membershipInfo?: string | null;
  /** @nullable */
  memberQuotes?: string | null;
  /** @nullable */
  promo?: string | null;
  /** @nullable */
  memberNotes?: string | null;
  isRecurring: boolean;
  /** @nullable */
  recurringPattern?: string | null;
  /**
     * ISO date string (YYYY-MM-DD) — recurrence ends after this date
     * @nullable
     */
  recurringUntil?: string | null;
  isGenderless: boolean;
  isPasswordProtected?: boolean;
  isDraft?: boolean;
  photoRequirement?: EventPhotoRequirement;
  createdAt: string;
  updatedAt: string;
}

export type EventInputPhotoRequirement = typeof EventInputPhotoRequirement[keyof typeof EventInputPhotoRequirement];


export const EventInputPhotoRequirement = {
  none: 'none',
  optional: 'optional',
  required: 'required',
} as const;

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  /** Time string (HH:MM) */
  time: string;
  dresscode?: string;
  category?: string;
  imageUrl?: string;
  registrationUrl?: string;
  tickettailorEmbed?: string;
  areaDescription?: string;
  membershipInfo?: string;
  memberQuotes?: string;
  promo?: string;
  memberNotes?: string;
  isRecurring?: boolean;
  recurringPattern?: string;
  recurringUntil?: string;
  isGenderless?: boolean;
  isPasswordProtected?: boolean;
  password?: string;
  isDraft?: boolean;
  photoRequirement?: EventInputPhotoRequirement;
}

export type EventUpdatePhotoRequirement = typeof EventUpdatePhotoRequirement[keyof typeof EventUpdatePhotoRequirement];


export const EventUpdatePhotoRequirement = {
  none: 'none',
  optional: 'optional',
  required: 'required',
} as const;

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  date?: string;
  time?: string;
  dresscode?: string;
  category?: string;
  imageUrl?: string;
  registrationUrl?: string;
  tickettailorEmbed?: string;
  areaDescription?: string;
  membershipInfo?: string;
  memberQuotes?: string;
  promo?: string;
  memberNotes?: string;
  isRecurring?: boolean;
  recurringPattern?: string;
  recurringUntil?: string;
  isGenderless?: boolean;
  isPasswordProtected?: boolean;
  password?: string;
  isDraft?: boolean;
  photoRequirement?: EventUpdatePhotoRequirement;
}

export interface ParticipationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  contact: string;
  photoUrl?: string;
}

export interface ParticipationResponse {
  id: number;
  eventId: number;
  name: string;
  contact: string;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface GalleryPhoto {
  id: number;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface GalleryPhotoInput {
  /** @minLength 1 */
  imageUrl: string;
  caption?: string;
}

export interface Banner {
  id: number;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface BannerInput {
  /** @minLength 1 */
  imageUrl: string;
  caption?: string;
  sortOrder?: number;
}

export type ProfileMemberType = typeof ProfileMemberType[keyof typeof ProfileMemberType];


export const ProfileMemberType = {
  singolo: 'singolo',
  coppia: 'coppia',
  singola: 'singola',
  trav: 'trav',
} as const;

export type ProfileInterestsItem = typeof ProfileInterestsItem[keyof typeof ProfileInterestsItem];


export const ProfileInterestsItem = {
  swinger: 'swinger',
  sexpositive: 'sexpositive',
  kinky: 'kinky',
  gangbang: 'gangbang',
} as const;

export interface Profile {
  id: number;
  nickname: string;
  age: number;
  email: string;
  telegram?: string | null;
  whatsapp?: string | null;
  memberType: ProfileMemberType;
  interests: ProfileInterestsItem[];
  createdAt: string;
  updatedAt: string;
}

export type ProfileInputMemberType = typeof ProfileInputMemberType[keyof typeof ProfileInputMemberType];


export const ProfileInputMemberType = {
  singolo: 'singolo',
  coppia: 'coppia',
  singola: 'singola',
  trav: 'trav',
} as const;

export type ProfileInputInterestsItem = typeof ProfileInputInterestsItem[keyof typeof ProfileInputInterestsItem];


export const ProfileInputInterestsItem = {
  swinger: 'swinger',
  sexpositive: 'sexpositive',
  kinky: 'kinky',
  gangbang: 'gangbang',
} as const;

export interface ProfileInput {
  /** @minLength 1 */
  nickname: string;
  /** @minimum 18 */
  age: number;
  /** @minLength 3 */
  email: string;
  telegram?: string;
  whatsapp?: string;
  memberType: ProfileInputMemberType;
  interests: ProfileInputInterestsItem[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorResponse {
  error: string;
}