# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      description: Returns all events sorted by date ascending
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/participate:
    post:
      operationId: participateInEvent
      tags: [events]
      summary: Register interest in an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/participations:
    get:
      operationId: listParticipations
      tags: [events]
      summary: List participations for an event (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of participations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipationResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gallery:
    get:
      operationId: listGalleryPhotos
      tags: [gallery]
      summary: List all gallery photos
      responses:
        "200":
          description: List of gallery photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPhoto"
    post:
      operationId: createGalleryPhoto
      tags: [gallery]
      summary: Add a gallery photo (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryPhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryPhoto
      tags: [gallery]
      summary: Delete a gallery photo (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List all communication banners
      responses:
        "200":
          description: List of banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Add a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /banners/{id}:
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, title, description, date, time, isRecurring, isGenderless, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
          description: ISO date string (YYYY-MM-DD)
        time:
          type: string
          description: Time string (HH:MM)
        dresscode:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        registrationUrl:
          type: ["string", "null"]
        tickettailorEmbed:
          type: ["string", "null"]
        areaDescription:
          type: ["string", "null"]
        membershipInfo:
          type: ["string", "null"]
        memberQuotes:
          type: ["string", "null"]
        promo:
          type: ["string", "null"]
        memberNotes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        recurringPattern:
          type: ["string", "null"]
        recurringUntil:
          type: ["string", "null"]
          description: ISO date string (YYYY-MM-DD) — recurrence ends after this date
        isGenderless:
          type: boolean
        isPasswordProtected:
          type: boolean
        isDraft:
          type: boolean
        photoRequirement:
          type: string
          enum: [none, optional, required]
        createdAt:
          type: string
        updatedAt:
          type: string
    EventInput:
      type: object
      required: [title, date, time]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
          description: ISO date string (YYYY-MM-DD)
        time:
          type: string
          description: Time string (HH:MM)
        dresscode:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        registrationUrl:
          type: string
        tickettailorEmbed:
          type: string
        areaDescription:
          type: string
        membershipInfo:
          type: string
        memberQuotes:
          type: string
        promo:
          type: string
        memberNotes:
          type: string
        isRecurring:
          type: boolean
        recurringPattern:
          type: string
        recurringUntil:
          type: string
        isGenderless:
          type: boolean
        isPasswordProtected:
          type: boolean
        password:
          type: string
        isDraft:
          type: boolean
        photoRequirement:
          type: string
          enum: [none, optional, required]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
        time:
          type: string
        dresscode:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        registrationUrl:
          type: string
        tickettailorEmbed:
          type: string
        areaDescription:
          type: string
        membershipInfo:
          type: string
        memberQuotes:
          type: string
        promo:
          type: string
        memberNotes:
          type: string
        isRecurring:
          type: boolean
        recurringPattern:
          type: string
        recurringUntil:
          type: string
        isGenderless:
          type: boolean
        isPasswordProtected:
          type: boolean
        password:
          type: string
        isDraft:
          type: boolean
        photoRequirement:
          type: string
          enum: [none, optional, required]
    ParticipationInput:
      type: object
      required: [name, contact]
      properties:
        name:
          type: string
          minLength: 1
        contact:
          type: string
          minLength: 1
        photoUrl:
          type: string
    ParticipationResponse:
      type: object
      required: [id, eventId, name, contact, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        contact:
          type: string
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    GalleryPhoto:
      type: object
      required: [id, imageUrl, sortOrder, createdAt]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    GalleryPhotoInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          minLength: 1
        caption:
          type: string
    Banner:
      type: object
      required: [id, imageUrl, sortOrder, createdAt]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    BannerInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          minLength: 1
        caption:
          type: string
        sortOrder:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string